+++ extensions/punch-clock/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Punch Clock Changelog
 
+## [Menu Bar Setup Check] - 2026-09-20
+
+- Warn when starting a timer if the "Work Timer" menu bar command has never
+  run, since the timer's countdown won't be visible until it's enabled
+
 ## [Initial Version] - 2026-09-11
 
 - Enter your total working time and break length for the day